Add files via upload
This tidy data set contains 4,898 white wines with 11 variables on quantifying the chemical properties of each wine. At least 3 wine experts rated the quality of each wine, providing a rating between 0 (very bad) and 10 (very excellent).

The main feature of interest in the dataset is **quality**. This project is to explore which chemical properties influence the quality of white wines.